Refactor Home Page
